extract boxoffice-mojo to module
