+++ lib/movie_wager_backend/application.ex
@@ -1,30 +1,19 @@
 defmodule MovieWagerBackend.Application do
   use Application
 
-  # See https://hexdocs.pm/elixir/Application.html
-  # for more information on OTP Applications
   def start(_type, _args) do
     import Supervisor.Spec
 
-    # Define workers and child supervisors to be supervised
     children = [
-      # Start the Ecto repository
       supervisor(MovieWagerBackend.Repo, []),
-      # Start the endpoint when the application starts
       supervisor(MovieWagerBackendWeb.Endpoint, []),
-      # Start your own worker by calling: MovieWagerBackend.Worker.start_link(arg1, arg2, arg3)
-      # worker(MovieWagerBackend.Worker, [arg1, arg2, arg3]),
-      worker(MovieWagerBackend.BoxOfficeCollector, []),
+      worker(MovieWagerBackend.BoxOfficeCollector, [])
     ]
 
-    # See https://hexdocs.pm/elixir/Supervisor.html
-    # for other strategies and supported options
     opts = [strategy: :one_for_one, name: MovieWagerBackend.Supervisor]
     Supervisor.start_link(children, opts)
   end
 
-  # Tell Phoenix to update the endpoint configuration
-  # whenever the application is updated.
   def config_change(changed, _new, removed) do
     MovieWagerBackendWeb.Endpoint.config_change(changed, removed)
     :ok
